% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.equal.R
\name{all.equal}
\alias{all.equal}
\alias{all.equal.DynMdl}
\title{Test if two \code{\link{DynMdl}} objects are 
(nearly) equal}
\usage{
\method{all.equal}{DynMdl}(target, current, ...)
}
\arguments{
\item{target}{and \code{DynMdl} (or \code{\link{FitMdl}}) object}

\item{current}{another \code{DynMdl} object, to be compared with
\code{target}}

\item{...}{Arguments passed to the internal call of
\code{\link[base]{all.equal}}.}
}
\value{
Either \code{TRUE} or a character vector describing the
differences between target and current.
}
\description{
\code{all.equal(x, y)} is a utility to compare R objects x and y testing
near equality. If they are different, comparison is still made to some
extent, and a report of the differences is returned. Do not use
\code{all.equal} directly in if expressions - use
\code{isTRUE(all.equal(...))}.
}
\details{
The implementation of \code{all.equal} for \code{DynMdl} objects
first serialized the model using the \code{DynMdl} method
\code{serialize} and then uses \code{\link[base]{all.equal}} of
the base package.
}
\examples{
mdl <- islm_mdl("2017Q2/2018Q2")
mdl2 <- mdl$copy()
print(all.equal(mdl, mdl2))

# now modify mdl2
mdl2$set_endo_values(600, names = "c")
print(all.equal(mdl, mdl2))
}
\seealso{
\code{\link[base]{all.equal}}
}
