% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{check_dynare}
\alias{check_dynare}
\title{\code{\link{DynMdl}} method: Compute the eigenvalues of the linearized model
around the steady state with Dynare}
\description{
Compute the eigenvalues of the linearized model around the steady state with
Dynare using Matlab or Octave.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
mdl$check_dynare(scratch_dir = tempfile(), dynare_path = NULL,
                  model_options = list(), 
                  use_octave = Sys.which("matlab") == "",
                  exit_matlab = FALSE)
}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{

\describe{
\item{\code{scratch_dir}}{Directory where the Matlablab and Dynare scripts are
created.  By default this is a temporary directory that is automatically
deleted when the R session terminates.}
\item{\code{dynare_path}}{Character string specifying the name of the
directory of the Dynare installation. On Linux it is usually not necessary
to the specify this argument. On Windows it is necessary to specify the path
of the Dynare installation. In you are running R in the  CPB
environment the path to Dynare is set automatically.}
\item{\code{model_options}}{Options passed to the \code{model} command of
Dynare. This should be a named list, which names corresponding to the Dynare
options. Specify a \code{NULL} value if the option has no value.
Consult the documentation of  Dynare for a list of available options.
Example: \code{model_options = list(block = NULL, mfs = 2)}}
\item{\code{use_octave}}{A logical. If \code{TRUE}, then
Dynare is envoked with Octave, otherwise Matlab is used. By default
Matlab is used if available.}
\item{\code{exit_matlab}}{A logical specifying if Matlab
should immediately  exit when the calcultions have finished
Matlab writes the output to a separate console. If \code{exit_matlab} is
\code{FALSE} (the default), then the R job waits until the user has closed
this console, or entered \code{exit} in the console. Otherwise the console
is automatically closed at the end of the calculation and all output is lost.
This argument is ignored if Dynare is run with Octave. Octave does not
open a separate console: all output appears in the same console used by R.}
}
}

\examples{
\dontrun{
islm <- islm_mdl()
islm$check_dynare()
print(islm$get_eigval())
}
}
\seealso{
\code{\link{check}},  \code{\link{solve_steady_dynare}}
and \code{\link{solve_dynare}}.
}
