% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMdl_doc.R
\name{clear_fit}
\alias{clear_fit}
\title{\code{\link{FitMdl}} method: deletes all fit targets}
\description{
This methods of R6 class \code{\link{FitMdl}}
deletes all fit targets for the fit procedure.
}
\section{Usage}{

\preformatted{
mdl$clear_fit()
}
\code{mdl} is an \code{\link{FitMdl}} object
}

\seealso{
\code{\link{set_fit}} and \code{\link{get_fit}}
}
