% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model}
\alias{compile_model}
\title{Compile a model}
\usage{
compile_model(mod_file)
}
\arguments{
\item{mod_file}{the name of the model file}
}
\value{
an integer
}
\description{
Compile a model
}

