% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_first_order.R
\name{compute_first_order}
\alias{compute_first_order}
\title{Computes the first order conditions of a model file}
\usage{
compute_first_order(mod_file)
}
\arguments{
\item{mod_file}{the filename of the mod file}
}
\value{
\code{NULL}
}
\description{
Computes the first order conditions of a model file
}

