% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_example_mod.R
\name{copy_example_mod}
\alias{copy_example_mod}
\title{Creates a copy of an example model file}
\usage{
copy_example_mod(model_name, filename = paste(model_name, "mod", sep = "."))
}
\arguments{
\item{model_name}{the name of the mod file (currently \code{"islm"} 
and \code{"ifn"} are supported)}

\item{filename}{the file name of the copy model file. By default, this
is \code{model_name} suffixed with extension \code{.mdl}.}
}
\value{
\code{TRUE} is the copy was succesful
}
\description{
The package \code{dynmod} includes a number of example model files
that can be used to get started with this package. These files are
located in the directory \code{models} of the package directory. This
function can be used to copy the model file from this directory to
a specified directory
}
\examples{
copy_example_mod("islm")
\dontshow{
dir.create("mod")
}
copy_example_mod("islm", filename = "mod/islm.mod")
\dontshow{
unlink("islm.*")
unlink("mod", recursive = TRUE)
}
}
\seealso{
\code{\link{create_mod}} and \code{\link{DynMod}}
}
