% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_mdl.R
\name{dyn_mdl}
\alias{dyn_mdl}
\title{Creates a \code{\link{DynMdl}} or  \code{\link{FitMdl}} object from a mod file}
\usage{
dyn_mdl(mod_file, period, data, calc = c("R", "bytecode", "dll", "internal"),
  fit_mod_file, debug = FALSE, dll_dir, bytecode, use_dll)
}
\arguments{
\item{mod_file}{the name of the model file (including extension .mod)}

\item{period}{a \code{\link[regts]{period_range}} object}

\item{data}{the model data as a  \code{\link[regts]{regts}} object with column
names}

\item{calc}{Method used to evaluate the model equations. 
Possible values are \code{"R"}, \code{"bytecode"}, \code{"dll"} and 
\code{"internal"}. See details.}

\item{fit_mod_file}{the name of the generated fit mod file. If not specified,
then the fit mod file is destroyed after the model has been parsed.}

\item{debug}{If \code{TRUE}, then some print statements
are executed.}

\item{dll_dir}{the directory where the dynamically linked library is stored.
Primarily used for testing.
Only used if argument \code{use_dll} is \code{TRUE}.}

\item{bytecode}{Deprecated (replaced by argument \code{calc}. 
If \code{TRUE}, then the functions used to calculate the
residuals and Jacobian are byte compiled}

\item{use_dll}{Deprecated (replaced by argument \code{calc}. If \code{TRUE}, then the  functions used to calculate the
residuals and Jacobian are converted to C code that is subsequently
converted to a dynamically linked library.}
}
\value{
an \code{DynMdl} object or, if the mod file contains a
fit block, a \code{\link{FitMdl}} object.
}
\description{
Creates a \code{\link{DynMdl}} object from a mod file. If the mod file
contains a fit block, then this function returns a \code{\link{FitMdl}} 
object, which is an extension of a \code{DynMdl} object.
}
