% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMod.R
\docType{class}
\name{DynMod}
\alias{DynMod}
\title{An R6 class for a Dynare model}
\format{\code{\link{R6Class}} object.}
\usage{
DynMod
}
\value{
Object of \code{\link{R6Class}} containing a macro-economic model,
}
\description{
An R6 class for a Dynare model
}
\section{Fields}{

\describe{
\item{\code{endo_count}}{the number of endogenous variables}

\item{\code{endos}}{the names and steady values of the endogenous variables. These
values are initialised with the values from the \code{initval} block of the
mod-file and may be computed with the \code{steady} method}

\item{\code{exos}}{the names and steady values of the exogenous variables. These
values are initialised with the values from the \code{initval} block of the
mod-file}

\item{\code{params}}{the names and steady values of the parameters}

\item{\code{max_endo_lag}}{maximum lag for endogenous variables}

\item{\code{max_endo_lead}}{maximum lead for endogenous variables}

\item{\code{max_exo_lag}}{maximum lag for exogenous variables}

\item{\code{max_exo_lead}}{maximum lead for exogenous variables}

\item{\code{nstatic}}{the number of static variables (variables without lag or}

\item{\code{nfwrd}}{the number of forward looking variables (variables with}

\item{\code{npred}}{the number of backward looking variables (variables with}

\item{\code{nboth}}{the number of variables with both a lag and a lead}

\item{\code{ndynamic}}{the number of variables with a lag or lead}

\item{\code{nsfwrd}}{the number of variables with leads}

\item{\code{lead_lag_incidence}}{see documentation Dynare
of the model}

\item{\code{f_dynamic}}{function for the residuals and Jacobian for the dynamic version
of the model}

\item{\code{model_period}}{the model period as a \code{\link[regts]{regperiod_range}}}

\item{\code{endo_period}}{the data period for the endogenous variables}

\item{\code{exo_period}}{the data period for the exogenous variables}

\item{\code{endo_data}}{a \code{\link[regts]{regts}} with the values of the
endogenous variables}

\item{\code{exo_data}}{a \code{\link[regts]{regts}} with the values of the
exogenous variables}

\item{\code{solve_out}}{the output of the solver (e.g. \code{\link[nleqslv]{nleqslv}})
used to solve the model}
}}
\section{Methods}{

\describe{

\item{\code{set_period(period)}}{Sets the model period. \code{period}
is a \code{\link{regperiod_range}} object or an object that can be coerced
to a \code{regperiod_range}. The model period is the longest period for which
the model may be solved. This method also allocates storage for
all model timeseries.  Model timeseries are
available for the so called 'model data period', which is
the model period extended with a lag and lead period. This method
also initialises all model timeseries with steady state values stored in the
\code{endos} and \code{exos} fields.}

\item{\code{solve_steady()}}{Solve the steady state of the model.
This methods solves the steady state problem. The data in field \code{endos}
is used as initial values for the endogenous variables.
This field is overwritten with the computed steady state values}

\item{\code{check()}}{Compute the eigenvalues of the linear
system and check if the Blachard and Kahn conditions are satisfied.}

\item{\code{solve(control = list())}}{Solves the model using a stacked-time
Newton method for the whole model period.
Argument \code{control} is a list with solve options (TODO: describe these
options somewhere).}

\item{\code{solve_perturbation()}}{Solves the model using the perturbation
theory used in the Dynare function stoch_simul. Only shocks in the first
solution period are allowed.}

\item{\code{get_jacob(sparse = TRUE)}}{Returns the Jacobian for the
stacked-time Newton problem either as a sparse matrix
(a \code{\link[Matrix]{Matrix}} object) or normal \code{\link{matrix}}.}

}
}
\keyword{data}

