% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMod.R
\docType{class}
\name{FitMod}
\alias{FitMod}
\title{An R6 class for a Dynare model with Fit targets
This class is a subclass of a \code{\link{DynMod}} objects. It contains
special methods for the fit procedure.}
\format{\code{\link{R6Class}} object.}
\usage{
FitMod
}
\value{
Object of \code{\link{R6Class}} containing a macro-economic model,
}
\description{
An R6 class for a Dynare model with Fit targets
This class is a subclass of a \code{\link{DynMod}} objects. It contains
special methods for the fit procedure.
}
\section{Methods}{

\describe{

\item{\code{get_endo_names()}}{Returns the names of the endogenous variables
of the model (excluding the residuals and Lagrange multipliers used in the
fit procedure).}

\item{\code{get_exo_names()}}{Returns the names of the exogenous variables
of the model (excluding the fit control variables).}

\item{\code{set_fit_targets(data, names = colnames(data))}}{
Sets the targets for the fit procedure. \code{data} is a
\code{regts} or \code{ts}. With argument \code{names} the names of the
timeseries in \code{data} can be specified. This argument
is mandatory if \code{data} does not have column names.
Only non \code{NA} values are registered as fit target values. 
A fit target can be removed by specifying a \code{NA}.}

\item{\code{get_fit_targets())}}{Returns the fit targets
used in the fit procedure.}

\item{\code{get_residuals())}}{Returns the residuals used in the 
fit procedure.}

\item{\code{get_lagrange())}}{Returns the Lagrange multipliers
used in the fit procedure.}

\item{\code{get_exo_data(names, period = self$get_data_period()}}{Returns
the exogenous data excluding the control variables of the
fit procedure.
\code{pattern} is a regular expression,  \code{names} a list of variables
 and \code{period} an \code{\link[regts]{regperiod_range}} object
 or an object that can be coerced to \code{regperiod_range}.}

\item{\code{get_endo_data(pattern, names, period = self$get_data_period()}}{Returns
the endogenous data excluding the auxiliary endogenous variables
 for the fit procedure (the residuals and the Lagrange multipliers).
\code{pattern} is a regular expression,  \code{names} a list of variables
 and \code{period} an \code{\link[regts]{regperiod_range}} object
 or an object that can be coerced to \code{regperiod_range}.}
}
}

\keyword{data}
