% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_data-methods}
\alias{get_data-methods}
\alias{get_endo_data}
\alias{get_exo_data}
\title{\code{\link{DynMdl}} methods: Retrieve timeseries from the model data}
\description{
These methods of R6 class \code{\link{DynMdl}} 
can be used to retrieve timeseries from the model data.
}
\section{Usage}{

\preformatted{
mdl$get_endo_data(pattern, names, period = mdl$get_data_period())

mdl$get_exo_data(pattern, names, period = mdl$get_data_period())

}

\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{names}}{a character vector with variable names}
\item{\code{pattern}}{a regular expression}
\item{\code{period}}{an \code{\link[regts]{period_range}} object or an
object that can be coerced to a \code{period_range}}
}
}

\section{Methods}{

\itemize{
\item \code{get_endo_data}: Endogenous model variables

\item \code{get_exo_data}: Exogenous model variables
}
}

