% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_eigval}
\alias{get_eigval}
\title{\code{\link{DynMdl}} method: Return the eigenvalues computed with method
\code{check}}
\description{
This method of R6 class \code{\link{DynMdl}} returns the eigenvalues 
computed with method \code{\link{check}}, ordered with increasing absolute 
value
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
mdl$get_eigval()
}

\code{mdl} is an \code{\link{DynMdl}} object
}

\seealso{
\code{\link{check}}
}
