% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_equations}
\alias{get_equations}
\alias{get_original_equations}
\alias{get_static_equations}
\title{\code{\link{DynMdl}} method: Returns a character vector with the
dynamic model equations}
\description{
These method of R6 class \code{\link{DynMdl}} returns a character vector
with the model equations (excluding local equations) of the static or
dynamic version of the model.  \code{get_static_equations} and
\code{get_equations} return the parsed equations of the static and dynamic model,
respectively. For models with trends, these function return the detrended
equations, the equations that are actually used when solving the model.
\code{get_original_equations} returns the equations as defined in the mod
file.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
 mdl$get_static_equations(i)
 mdl$get_equations(i)
 mdl$get_original_equations(i)
}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{i}}{A numeric vector with the indices of the non-local equations.
If not specified, then the function returns all equations}
}
}

\examples{
mdl <- islm_mdl(period = "2018Q1/2023Q3")

# print the 4th equation nicely to the screen
cat(mdl$get_equations(4))

# print all equations
print(mdl$get_equations())
}
