% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMdl_doc.R
\name{get_fit-methods}
\alias{get_fit-methods}
\alias{get_fit}
\alias{get_fit_instruments}
\alias{get_lagrange}
\title{\code{\link{FitMdl}} methods: get variables used in the fit procedure.}
\description{
These methods of R6 class \code{FitMdl} can be used
to retrieve the variables used in the fit procedure:
the fit targets, fit instruments or Lagrange multipliers.

For method \code{get_fit} there are corresponding \code{\link{set_fit}} and
\code{\link{set_fit_values}} methods. There are currently no special
methods to set or change the fit instruments and Lagrange multipliers.
However, since they are internally implemented as endogenous variables
you can use methods \code{\link{set_data}}, \code{\link{set_endo_values}},
and \code{\link{change_endo_data}} to change the fit instruments
or Lagrange multipliers.
}
\section{Usage}{

\preformatted{

mdl$get_fit() # fit targets

mdl$get_fit_instruments(pattern, names, period = mdl$get_period())

mdl$get_lagrange(names, period = mdl$get_period())
}

\code{mdl} is an \code{\link{FitMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression}
\item{\code{names}}{a character vector with variable names}
\item{\code{period}}{an \code{\link[regts]{period_range}} object or an
object that can be coerced to a \code{period_range}}
}
}

\examples{

mdl <- islm_mdl(period = "2016Q1/2017Q3", fit = TRUE)

# create a regts with fit targets
y <- regts(c(1250, 1255, 1260), start = "2016Q1")
t <- regts(c(250, 255), start = "2016Q1")
fit_targets <- cbind(y, t)

# register the fit targets in the FitMdl object
mdl$set_fit(fit_targets)

mdl$solve()

print(mdl$get_fit())
print(mdl$get_fit_instruments())
print(mdl$get_lagrange())

}
\seealso{
\code{\link{get_data-methods}}, \code{\link{set_fit}} and  
\code{\link{set_fit_values}}.
}
