% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc_fit.R
\name{get_fit_steady}
\alias{get_fit_steady}
\title{\code{\link{DynMdl}} method: get fit targets used in the steady state
calculation}
\description{
This methods of R6 class \code{DynMdl} returns the static fit targets,
i.e. the fit targets used when the steady state is solved.
\cr\cr
By default, function \code{get_fit} only returns fit targets with
any non-\code{NA} value.
If all values are \code{NA}, then the function returns \code{NULL}.
If \code{names} or \code{pattern} has been specified, it always returns a timeseries
with the specified variables.
\cr\cr
For method \code{get_fit_syteady} the is a corresponding \code{\link{set_fit}_steady}
method. There are currently no special
methods to set or change the fit instruments and Lagrange multipliers.
However, since they are internally implemented as endogenous variables
you can use methods \code{\link{set_static_data}}  to change the static
fit instruments or Lagrange multipliers.
}
\section{Usage}{

\preformatted{

mdl$get_fit_steady(pattern, names) # fit targets

}

\code{mdl} is a \code{\link{DynMdl}} object implementing the fit method.
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression for selecting the names of
the fit targets.}
\item{\code{names}}{a character vector with variable names.}
}
}

\examples{

mdl <- islm_mdl(fit = TRUE)

fit_targets <- c(y = 1250, t = 255)

# register the fit targets in the DynMdl object
mdl$set_fit_steady(fit_targets)

mdl$solve_steady()

print(mdl$get_fit_steady())

}
\seealso{
\code{\link{set/get_static_endos/exos}},
\code{\link{set_fit_steady}},
and \code{\link{clear_fit}}.
}
