% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc_fit.R
\name{get_instrument_names/get_sigma_names}
\alias{get_instrument_names/get_sigma_names}
\alias{get_instrument_names}
\alias{get_sigma_names}
\title{\code{\link{DynMdl}} methods: Retrieve the names of the fit instruments
or sigma parameters used in the fit method.}
\description{
These methods of R6 class \code{\link{DynMdl}}
return the names of the fit instruments or sigma parameters
used in the fit method.
}
\section{Usage}{

\preformatted{
mdl$get_instrument_names(all = FALSE)

mdl$get_sigma_names()

}
\code{mdl} is a \code{\link{DynMdl}} object implementing the fit method.
}

\section{Arguments}{

\describe{
\item{\code{all}}{A logical (default \code{FALSE}). If \code{TRUE}, the
names of all fit instruments are returned, including the inactive
fit instruments.}
}
}

\examples{

mdl <- islm_mdl(period = "2017Q1/2018Q3", fit = TRUE)
print(mdl$get_instrument_names())
print(mdl$get_sigma_names())

}
\seealso{
\code{\link{get_fit_instruments}}, \code{\link{get_sigmas}}
}
