% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacob.R
\docType{methods}
\name{get_jacob}
\alias{get_jacob}
\alias{get_jacob,SimModel-method}
\title{Returns the dense stacked-time Jacobian}
\usage{
get_jacob(mdl, sparse = TRUE)

\S4method{get_jacob}{SimModel}(mdl, sparse = TRUE)
}
\arguments{
\item{mdl}{a \code{\link{SimModel}}}

\item{sparse}{If \code{TRUE} then the function returns a dense Jacobiab,
otherwise a sparse Jacobian created with the  a \code{\link{Matrix}}
package}
}
\value{
the stacked-time Jacobian
}
\description{
This function returns the stacked-time Jacobian for the complete
solve period as a \code{\link{Matrix}} or \code{\link{matrix}}
(depending on argument \code{sparse})
}
\section{Methods (by class)}{
\itemize{
\item \code{SimModel}: Returns the dense stacked-time Jacobian
for the solution period
}}

