% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_max_lag/get_max_lead}
\alias{get_max_lag/get_max_lead}
\alias{get_max_lag}
\alias{get_max_lead}
\title{\code{\link{DynMdl}} methods: Returns the maximum lag or lead of the model}
\description{
Methods \code{get_max_lag} and \code{get_max_lead} of R6 class 
\code{\link{DynMdl}} return the maximum lag and lead, respectively.
 the maximum a character vector
}
\section{Usage}{

\code{DynMdl} methods:
\preformatted{
 mdl$get_max_lag()
 mdl$get_max_lead()
 
}

\code{mdl} is an \code{\link{DynMdl}} object
}

