% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_mdldef}
\alias{get_mdldef}
\title{\code{\link{DynMdl}} method: Returns technical details about the model.}
\description{
This function returns a list with various components containing technical
details of the model, such as the names of the variables,
the lead-lag incidence matrix, etc.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
mdl$sget_mdldef()
}

\code{mdl} is a \code{\link{DynMdl}} object
}

