% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc_fit.R
\name{get_sigmas}
\alias{get_sigmas}
\title{\code{\link{DynMdl}} methods: Retrieve the
sigma parameters used in the fit method.}
\description{
This method of R6 class \code{\link{DynMdl}}
return all sigma parameters for the fit method larger than or
equal to zero. The fit instruments for which the corresponding
sigma-parameter is \eqn{\ge 0} are used as instruments in the
fit method. Instruments which a sigma-parameter smaller than 0 are
fixed at their original value.

The sigma-parameters can be modified with methods
\code{\link{set_param}} and \code{\link{set_param_values}}.
}
\section{Usage}{

\preformatted{
mdl$get_sigmas()

\code{mdl} is a \code{\link{DynMdl}} object implementing the fit method.
}
}

\examples{

mdl <- islm_mdl(period = "2016Q1/2017Q3", fit = TRUE)
mdl$get_sigmas()

# disable ui as fit instrument by setting sigma_ui to -1:
mdl$set_param(c(sigma_ui = -1))
mdl$get_sigmas()
}
