% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_solve_status}
\alias{get_solve_status}
\title{\code{\link{DynMdl}} method: Returns the solve status of the last model solve.}
\description{
This method of R6 class \code{\link{DynMdl}} returns the status
of the last model solve as a text string. If the last model solve
was succesfull, it returns the string \code{"OK"}.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
mdl$get_solve_status()

}
\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Details}{


The possible return values are:
\itemize{
\item \code{NA_character_} (method solve has not yet been called)
\item \code{"OK"}
\item \code{"ERROR"} (an error has occurred, check the warnings).
}
}

\examples{
\dontrun{
mdl <- islm_mdl(period = "2017Q1/2018Q4")
mdl$set_endo_values(NA, names = "y", period = "2017Q1")
mdl$solve()
if (mdl$get_solve_status() != "OK") {
   stop("Error solving the model. Check the warnings!")
}
}
}
\seealso{
\code{\link{solve} and \code{\link{solve_steady}}}
}
