% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islm_mdl.R
\name{islm_mdl}
\alias{islm_mdl}
\title{Returns an example ISLM model}
\usage{
islm_mdl(period = NULL)
}
\arguments{
\item{period}{the model period for the ISLM model}
}
\value{
a \code{\link{DynMdl}} object
}
\description{
This function returns an example ISLM model, If argument \code{period} has
been specified, then this function also initializes the model data
with the steady state values.
}
\examples{
mdl <- islm_mdl("2017Q1/2019Q4")
}
