% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_period.R
\docType{methods}
\name{model_period}
\alias{model_period}
\alias{model_period,SimModel-method}
\alias{model_period<-}
\alias{model_period<-,SimModel-method}
\title{Get or set the solve period of model object}
\usage{
model_period(mdl)

model_period(mdl) <- value

\S4method{model_period}{SimModel}(mdl)

\S4method{model_period}{SimModel}(mdl) <- value
}
\arguments{
\item{mdl}{the model}

\item{value}{a \code{\link[regts]{regperiod_range}} representing the solve period}
}
\value{
a \code{\link[regts]{regperiod_range}} representing the solve period
}
\description{
Get or set the solve period of model object
}
\section{Methods (by class)}{
\itemize{
\item \code{SimModel}: Returns the solve period

\item \code{SimModel}: Sets the solve period of a  \code{\link{SimModel}} object
}}

