% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{put_static_endos}
\alias{put_static_endos}
\title{\code{\link{DynMdl}} method: Transfers the static endogenous
variables to the model data.}
\description{
This method of R6 class \code{\link{DynMdl}} transfers the static endogenous
variables to the model data.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
mdl$put_static_endos(period = mdl$get_data_period())

}

\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{period}}{
A \code{\link[regts]{period_range}} object or an object that can be 
coerced to a \code{period_range}}, specifying the period for which the 
endogenous model data will be updated with the computed steady state values.
}
}

\examples{
mdl <- islm_mdl()

# transfer static endogenous variables for the full data period 
mdl$put_static_endos()

# now only for the lead period
mdl$put_static_endos(period = mdl$get_lead_period()))
}
\seealso{
\code{\link{solve_steady}}, \code{\link{set_static_endsos}} 
and \code{\link{get_static_endos}}.
}
