% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mdl.R
\name{read_mdl}
\alias{read_mdl}
\title{Reads a model from a RDS file}
\usage{
read_mdl(file, dll_dir)
}
\arguments{
\item{file}{the name of the RDS file}

\item{dll_dir}{the directory where the dynamically linked library is stored.
Primarily used for testing.
Only used if argument \code{use_dll} was \code{TRUE} (see function
\code{\link{create_mdl}}.}
}
\value{
a \code{\link{DynMdl}} or \code{\link{FitMdl}} object
}
\description{
This function reads a model from an RDS file that has been written
by method \code{\link{write_mdl}} of an \code{\link{DynMdl}} or
\code{\link{FitMdl}} object.
}
\examples{
mdl <- islm_mod("2017Q1/2019Q2")
mdl$write_mdl("islm_mod.rds")
mdl2 <- read_mdl("islm_mod.rds")
}
\seealso{
\code{\link{write_mod}}
}

