% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{set_data}
\alias{set_data}
\title{\code{\link{DynMdl}} method: transfers data from a timeseries 
object to the model data}
\description{
This method of R6 class \code{\link{DynMdl}} 
transfers data from a timeseries object to the model data
(both endogenous and exogenous)
}
\section{Usage}{

\preformatted{
mdl$set_data(data, names = colnames(data), 
             update_mode = c("update", "updval")

}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{data}}{a \code{\link[stats]{ts}} or \code{\link[regts]{regts}}
 object}
\item{\code{names}}{a character vector with variable names. Defaults to the
column names of \code{data}. If \code{data} does not have column names,
then argument \code{names} is mandatory}
\item{\code{update_mode}}{the update mode. See details.}
}
}

\section{Details}{


The argument \code{update_mode} controls how the timeseries are updated:
\describe{
\item{\code{"update"}}{Model variables are updated with the 
timeseries in \code{data}}
\item{\code{"updval"}}{Model variables are updated with the non \code{NA}
values in \code{data}}
}
}

