% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMdl_doc.R
\name{set_fit}
\alias{set_fit}
\title{\code{\link{FitMdl}} method: transfers data from a timeseries
object to the fit targets.}
\description{
The method \code{set_fit} of R6 class \code{\link{FitMdl}}
transfers data from a timeseries object to the fit targets.
}
\section{Usage}{

\preformatted{

mdl$set_fit(data, names = colnames(data), upd_mode = c("upd", "updval"))
}

\code{mdl} is an \code{\link{FitMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{data}}{a \code{\link[stats]{ts}} or \code{\link[regts]{regts}}
 timeseries object}
\item{\code{names}}{a character vector with variable names, with the
same length as the number of timeseries in \code{data}. Defaults to the
column names of \code{data}. If \code{data} does not have column names,
then argument \code{names} is mandatory}
\item{\code{upd_mode}}{the update mode, a character string specifying
how the timeseries are updated: \code{"upd"} (standard update, default) or
\code{"updval"} (update only with valid numbers). NOTE: currently
\code{"updval"} is not yet supported. See details.}
}
}

\section{Details}{


Method \code{set_fit} transfers data from a timeseries object to the
fit targets. It works similarly as method  \code{\link{set_data}}.
If \code{data} is a multivariate timeseries object, then
each column is used to update the fit target with the same
name as the column name. If \code{data} does not have column names,
or if the column names do not correspond to the model variable names,
then argument \code{names} should be specified.

By default, all values in \code{data} are used to update the corresponding
model variable. Sometimes it is desirable to skip the \code{NA} values
in \code{data}. This can be achieved by selecting \code{"updval"} for argument
\code{upd_mode}. Other non finite numbers (\code{NaN}, \code{Inf}, and
\code{-Inf}) are also disregarded for this update mode.
NOTE: currently, update mode \code{"updval"} has not yet been implemented.
}

\examples{

mdl <- islm_mdl(period = "2016Q1/2017Q3", fit = TRUE)

# create a regts with fit targets
y <- regts(c(1250, 1255, 1260), start = "2016Q1")
t <- regts(c(250, 255), start = "2016Q1")
fit_targets <- cbind(y, t)

# register the fit targets in the DynMdl object
mdl$set_fit(fit_targets)

print(mdl$get_fit())
}
\seealso{
\code{\link{get_fit}}
}
