% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{set_param}
\alias{set_param}
\title{\code{\link{DynMdl}} method: Sets the model parameters}
\description{
This method of R6 class \code{\link{DynMdl}}
sets the model parameters
}
\section{Usage}{

\preformatted{
mdl$set_param(p)

}
\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{p}}{a named numeric vector with parameter values.
The names are the names of the parameter}
}
}

\examples{
mdl <- islm_mdl()
mdl$set_param(c(i0 = 101))
}
