% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{set_period}
\alias{set_period}
\title{\code{\link{DynMdl}} method: sets the model period}
\description{
This method of R6 class \code{\link{DynMdl}} sets the model period.
This is the default period used when solving the model.

If the model data has not already been initialized with method
\code{\link{init_data}}, then \code{set_period} also initializes
the model data. In that case the model data period is set to the
specified model period extended with a lag and lead period. For models with
trends, the data period also includes the base period.
All endogenous and exogenous variables, fit instruments and lagrange
multipliers are set to the static values for the whole data period.
If static fit targets have been
specified with method \code{\link{set_fit_steady}}, then the dynamic fit
targets are set to the static fit targets for the whole data period.

If the model data has already been initialized with  method
\code{\link{init_data}}, then the new model period
should be compatible with the model data period.
In particular, the new model period extended with a lag and lead period
should not contain periods outside the model data period.
}
\section{Usage}{

\preformatted{
mdl$set_period(period)

}

\code{mdl} is a \code{DynMdl} object
}

\section{Arguments}{


\describe{
\item{\code{period}}{\code{\link[regts]{period_range}}
object, or an object that can be coerced to
\code{\link[regts]{period_range}}}
}
}

\examples{
mdl <- islm_mdl()
mdl$set_period("2017Q2/2021Q3")
}
