% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{set_static_values-methods}
\alias{set_static_values-methods}
\alias{set_static_endo_values}
\alias{set_static_exo_values}
\title{\code{\link{DynMdl}} methods: Sets the values of the static model data}
\description{
This method of R6 class \code{\link{DynMdl}}
can be used to set the static values of the model data.
}
\section{Usage}{

\preformatted{

mdl$set_static_endo_values(value, names, pattern)

mdl$set_static_exo_values(value, names, pattern)

}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{value}}{a numeric vector of length 1}
\item{\code{names}}{a character vector with variable names.}
\item{\code{pattern}}{a regular expression for selecting the names of variables.}
}
If neither \code{names} nor \code{pattern} have been specified,
then the static values of all endogenous or exogenous variables are set to the specified value.
}

\section{Methods}{

\itemize{
\item \code{set_static_endo_values}: Endogenous model variables
(including fit instruments and Lagrange multipliers for the fit method).

\item \code{set_static_exo_values}: Exogenous model variables.

}
}

\examples{

mdl <- islm_mdl()

# set the static value of ms
mdl$set_static_exo_values(205, names = "ms")

# set the static values for y and yd to 1000
mdl$set_static_endo_values(1000, pattern = "^yd?$")
}
\seealso{
\code{\link{set_static_data}} and
\code{\link{change_static_data-methods}} and
}
