% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{set_values-methods}
\alias{set_values-methods}
\alias{set_values}
\alias{set_endo_values}
\alias{set_exo_values}
\title{\code{\link{DynMdl}} methods: Sets the values of the model data}
\description{
This method of R6 class \code{\link{DynMdl}} 
can be used to set the values of the model dataents
(both endogenous and exogenous)
}
\section{Usage}{

\preformatted{
mdl$set_values(value, names, pattern, period = mdl$get_data_period())
mdl$set_endo_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_exo_values(value, names, pattern, period = mdl$get_data_period())

}

\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{value}}{a numeric vector of length 1 or with the same length
as the length of the range of \code{period}}
\item{\code{names}}{a character vector with variable names}
\item{\code{pattern}}{a regular expression}
\item{\code{period}}{a \code{\link[regts]{regperiod_range}} object or an
object that can be coerced to a \code{regperiod_range}}
}
}

\section{Methods}{

\itemize{
\item \code{set_values}: Endogenous and exogneous model variables

\item \code{set_endo_values}: Endogenous model variables

\item \code{set_exo_values}: Exogenous model variables

}
}

