% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimModel.R
\docType{class}
\name{SimModel-class}
\alias{SimModel}
\alias{SimModel-class}
\title{An S4 class for running simulations with the stacked-time Newton method}
\description{
An S4 class for running simulations with the stacked-time Newton method
}
\section{Slots}{

\describe{
\item{\code{exo_count}}{the number of exogeneous variables}

\item{\code{endo_count}}{the number of endogenous variables}

\item{\code{exos}}{the names and initial values of the exogenous variables}

\item{\code{endos}}{the names and initial values of the endogenous variables}

\item{\code{params}}{the names and values of the parameters}

\item{\code{max_endo_lag}}{maximum lag for endogenous variables}

\item{\code{max_endo_lead}}{maximum lead for endogenous variables}

\item{\code{max_exo_lag}}{maximum lag for exogenous variables}

\item{\code{max_exo_lead}}{maximum lead for exogenous variables}

\item{\code{lead_lag_incidence}}{see documentation Dynare}

\item{\code{f_static}}{function for the residuals and Jacobian for the static version
of the model}

\item{\code{f_dynamic}}{function for the residuals and Jacobian for the dynamic version
of the model}

\item{\code{model_period}}{the model period as a \code{\link[regts]{regperiod_range}}}

\item{\code{endo_period}}{the data period for the endogenous variables}

\item{\code{exo_period}}{the data period for the exogenous variables}

\item{\code{endo_data}}{a \code{\link[regts]{regts}} with the values of the
endogenous variables}

\item{\code{exo_data}}{a \code{\link[regts]{regts}} with the values of the
exogenous variables}

\item{\code{solve_out}}{the output of the solver (e.g. \code{\link[nleqslv]{nleqslv}}
used to solve the model}
}}

