% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_model.R
\docType{methods}
\name{solve_model}
\alias{solve_model}
\alias{solve_model,SimModel-method}
\title{Solve a \code{\link{SimModel}} object using the dense linear algebra.}
\usage{
solve_model(mdl, solver = c("sparse_newton", "nleqslv"), ...)

\S4method{solve_model}{SimModel}(mdl, solver = c("sparse_newton", "nleqslv",
  "dfsane", "BBsolve"), numjac = FALSE, ...)
}
\arguments{
\item{mdl}{a \code{SimModel}}

\item{solver}{the solve method}

\item{...}{arguments passed to the solver (TODO: add documentation about these
arguments, in particular for the sparse newton solver)}

\item{numjac}{If true, then the Jacobian will be calculated
numerically. Only used for the \code{nleqslv} solver}
}
\value{
a new \code{SimModel} with the solved model
}
\description{
The method is using various solvers (e.g. \code{\link[nleqslv]{nleqslv}}
for a dense linear algebra method).
}
\section{Methods (by class)}{
\itemize{
\item \code{SimModel}: Solve a  \code{\link{SimModel}}  model
using the dense linear algebra
}}

