% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_steady.R
\docType{methods}
\name{solve_steady}
\alias{solve_steady}
\alias{solve_steady,SimModel-method}
\title{Solve the steady state of a \code{\link{SimModel}} object}
\usage{
solve_steady(mdl)

\S4method{solve_steady}{SimModel}(mdl)
}
\arguments{
\item{mdl}{a \code{SimModel}}
}
\value{
the return a \code{SimModel} object with the \code{endos} slot set to
the steady state values
}
\description{
The steady state system is solved with the package \code{\link[nleqslv]{nleqslv}}
}
\section{Methods (by class)}{
\itemize{
\item \code{SimModel}: Solve the steady state of a \code{\link{SimModel}} object
}}

