% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{solve_steady_dynare}
\alias{solve_steady_dynare}
\title{\code{\link{DynMdl}} method: Solves the steady state with Dynare}
\description{
Solve the steady state and calculate the eigenvalues 
with Dynare using Matlab or Octave.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
mdl$solve_steady_dynare(scratch_dir = tempfile(), dynare_path = NULL,
                        model_options = list(), 
                        solve_options = list(tolf = 1e-8),
                        use_octave = Sys.which("matlab") == "")
}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{

\describe{
\item{\code{scratch_dir}}{Directory where the Matlablab and Dynare scripts are 
created.  By default this is a temporary directory that is automatically
deleted when the R session terminates.}
\item{\code{dynare_path}}{Character string specifying the name of the 
directory of the Dynare installation. On Linux it is usually not necessary 
to the specify this argument. On Windows it is necessary to specify the path 
of the Dynare installation.}
\item{\code{model_options}}{Options passed to the \code{model} command of
Dynare. This should be a named list, which names corresponding to the Dynare
options. Specify a \code{NULL} value if the option has no value.
Consult the documentation of  Dynare for a list of available options.
Example: \code{model_options = list(block = NULL, mfs = 2)}}
\item{\code{solve_options}}{Options passed to the 
\code{steady} command of Dynare. This should be a named list, which names 
corresponding to the Dynare options. Specify a \code{NULL} value if the 
option has no value. Consult the documentation of  Dynare for a list of 
available options.
Example: \code{steady_options = list(tolf = 1e-7, no_homotopy = NULL)}.
The default is \code{list(tolf = 1e-8)}}
\item{\code{use_octave}}{A logical. If \code{TRUE}, then
Dynare is envoked with Octave, otherwise Matlab is used. By default 
Matlab is used if available.}
}
}

\section{Value}{

 The eigenvalues of the steady state.
}

\examples{
\dontrun{
islm <- islm_mdl()
islm$solve_steady_dynare(solve_options = list(tolf = 1e-8))
}
}
