% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{write_initval_file}
\alias{write_initval_file}
\title{Writes the model data to a Dynare initval file}
\description{
This method of R6 class \code{\link{DynMdl}}
writes all endogous and exogenous model variables
to a so called "initval file" that can be read by Dynare.
An initval_file contains the paths of all model variables.
}
\section{Arguments}{


\describe{
\item{\code{file}}{the name of the initval file}
}
}

\examples{
mdl <- islm_mdl("2017Q1/2019Q2")

# write initval file as matlab file
mdl$write_initval_file("dynare_input/islm_initval.xlsx")

# write initval file in xlsx format
mdl$write_initval_file("dynare_input/islm_initval.xlsx")
}
