% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{change_data-methods}
\alias{change_data-methods}
\alias{change_data}
\alias{change_endo_data}
\alias{change_exo_data}
\title{\code{\link{DynMdl}} methods: changes the endogenous or exogenous
model data by applying a function.}
\description{
These methods of R6 class \code{\link{DynMdl}} changes endogenous and/or 
exogenous model data by applying a function.
}
\section{Usage}{

\preformatted{
mdl$change_endo_data(fun, names, pattern, period = mdl$get_data_period(), ...)

mdl$change_exo_data(fun names, pattern, period = mdl$get_data_period(), ...)

mdl$change_data(fun, names, pattern, period = mdl$get_data_period(), ...)
}

\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{fun}}{a function applied each model timeseries specified with 
 argument \code{names} or \code{pattern}}
\item{\code{names}}{a character vector with variable names}
\item{\code{pattern}}{a regular expression}
\item{\code{period}}{an \code{\link[regts]{period_range}} object or an
object that can be coerced to a \code{period_range}}
\item{\code{...}}{arguments passed to \code{fun}}
}
If neither \code{names} nor \code{pattern} have been specified,
then the function is applied to all endogenous or exogenous variables.
}

\section{Methods}{

\describe{
\item{\code{changes_endo_data}}{Changes endogenous model variables}
\item{\code{change_exo_data}}{Changes exogenous model variables}
\item{\code{change_data}}{Changes endogenous and/or exogenous model variables.
For \code{\link{FitMdl}} objects,
\code{change_data} can also be used to modify fit instruments and the  
Lagrange multipliers used in the fit method.}
}
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2017Q3")

# increase y and yd with 10\% for the full data period
mdl$change_endo_data(pattern = "^y.?$", fun = function(x) {x * 1.1})
print(mdl$get_endo_data())

# increase ms in 2017Q1 and 2017Q2 with 10 and 20, resp.
mdl$change_exo_data(names = "ms", fun = function(x, dx) {x + dx},
                dx = c(10, 20), period = "2017Q1/2017Q2")
print(mdl$get_exo_data())

}
\seealso{
\code{\link{get_data-methods}}, \code{\link{set_data}} and
\code{\link{set_values-methods}}
}
