% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mdl.R
\name{check_mdl}
\alias{check_mdl}
\title{This function computes the eigenvaues of the linearized model around the
steady state.}
\usage{
check_mdl(mod_file)
}
\arguments{
\item{mod_file}{the model file}
}
\description{
This function computes the eigenvaues of the linearized model around the
steady state.
}
