% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMdl_doc.R
\name{clear_fit}
\alias{clear_fit}
\title{\code{\link{FitMdl}} method: removes fit targets and turns off fit
instruments.}
\description{
This method of R6 class \code{FitMdl} removes all fit targets, sets 
the sigma-parameters of the fit-instruments to \code{-1} and sets
all Lagrange multipliers to 0. 

By removing the fit targets (which is equivalent to setting all fit targets to
\code{NA}), all endogenous variables are calculated according to the equations
of the model, while the fit instruments stay fixed at their current value, 
and are efficitively exogenous (even though they are still implemented
as endogenous variables).

If the model had been solved before \code{clear_fit} was called, then the
model is still solved after \code{clear_fit} has been called.
}
\section{Usage}{

\preformatted{

mdl$clear_fit()
}

\code{mdl} is an \code{\link{FitMdl}} object
'
}

\examples{

mdl <- islm_mdl(period = "2016Q1/2017Q3", fit = TRUE)

# create a regts with fit targets
y <- regts(c(1250, 1255, 1260), start = "2016Q1")
t <- regts(c(250, 255), start = "2016Q1")
fit_targets <- cbind(y, t)

# register the fit targets in the FitMdl object
mdl$set_fit(fit_targets)

mdl$solve()
mdl$clear_fit()

# the next statements gives 0 iterations.
mdl$solve()
}
\seealso{
\code{\link{get_data-methods}}, \code{\link{set_fit}},
\code{\link{set_fit_values}} and \code{\link{clear_fit}}.
}
