% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_mdl.R
\name{dyn_mdl}
\alias{dyn_mdl}
\title{Creates a \code{\link{DynMdl}} object from a mod file}
\usage{
dyn_mdl(
  mod_file,
  period,
  data,
  base_period = NULL,
  calc = c("internal", "R", "bytecode", "dll"),
  fit_mod_file,
  debug = FALSE,
  dll_dir,
  max_laglead_1 = FALSE,
  strict = TRUE,
  warn_uninit_param = TRUE,
  init_param_na = FALSE,
  fit = TRUE,
  fit_fixed_period = FALSE,
  check_static_eqs = TRUE,
  latex = TRUE,
  latex_options,
  nostrict,
  silent = FALSE
)
}
\arguments{
\item{mod_file}{the name of the model file (including extension .mod)}

\item{period}{a \code{\link[regts]{period_range}} object specifying the
model period, i.e. the period range for which the model will be solved. Thus
this period range excludes the lag and lead period.}

\item{data}{the model data as a  \code{\link[regts]{regts}} object with column
names}

\item{base_period}{a \code{\link[regts]{period}} object specifying
the base period for the trends. This is used if the model has trend variables.
All trend variables will be equal to 1 at the base period. This argument is
ignored for models without trend. If note specified  the \code{base_period} is
set to the start period of the model period.}

\item{calc}{Method used to evaluate the model equations.
Possible values are \code{"internal"},  \code{"R"}, \code{"bytecode"} and \code{"dll"}.
See Details.}

\item{fit_mod_file}{the name of the generated fit mod file. If not specified,
then the fit mod file is destroyed after the model has been parsed.
This argument should not be specified if the model contains
trends, since in that case the fit mod file cannot be used a input mod file
for function \code{dyn_mdl} or for Dynare. If wou want to check the
equations in the fit mod file, use argument \code{DEBUG} (see below).}

\item{debug}{If logical (default \code{FALSE}), only used when the
model is a fit model. If \code{TRUE}, then intermediate files created when
preparing the fit model are written to the current directory. By default
these files are written in a temporary directory and deleted when the
R session terminates.}

\item{dll_dir}{the directory where the dynamically linked library is stored.
Primarily used for testing.
Only used if argument \code{use_dll} is \code{TRUE}.}

\item{max_laglead_1}{a logical indicating whether the model should be
transformed internally to a model with a maximum lag and lead of 1.
The default is \code{FALSE}. This option has no effect if the maximum lag
and lead of the original model is 1. Set this argument to
\code{TRUE} if you want to analyse the stability of the steady state with
method \code{\link{check}} for models with a maximum lag or lead larger than 1.}

\item{strict}{A logical. If \code{TRUE} (the default), then an error
is given when endogenous or exogenous variables are not used in the model block
or when an undeclared symbol is assigned in the initval block. If
\code{FALSE}, then only a warning is issued,  unused endogenous variables
are removed, and the assignments of undeclared symbols in the initval block
are ignored.}

\item{warn_uninit_param}{A logical. If \code{TRUE} (the default) then
a warning is given for each parameter that has not been initialized in the
mod file. Uninitialized parameters are set to zero or \code{NA}, depending
on argument \code{init_param_na}.}

\item{init_param_na}{A logical (default \code{FALSE}). If \code{TRUE}, then
the parameters that have not been initialized in the
mod file are set to \code{NA}. Otherwise these parameters are initialised
with zero.}

\item{fit}{a logical (default \code{TRUE}) indicating if the \code{DynMdl}
object returned by this function should implement the fit procedure if the
mod file contains a fit block. Specify \code{FALSE} if the mod file  has a fit
block while you do not want to use the fit procedure.}

\item{fit_fixed_period}{a logical. If \code{TRUE}, then the fit conditions are
derived for a fixed period, treating lags and leads as exogenous variables.
If \code{FALSE} (the default), the fit conditions are derived from the
stacked-time equations. This option is particularly useful for backward
looking models (models without leads). If \code{fit_fixed_period} is \code{TRUE}, then
the fit equations will not contain lags or leads. If \code{fit_fixed_period} is
\code{TRUE}, then some fit equation may contain leads.}

\item{check_static_eqs}{a logical. If \code{TRUE} (the default), then we check
if the mod file contains separate static and dynamic equations (i.e.
equations tagged with \code{static} and \code{dynamic}). If this is the case,
separate static and dynamic fit equations are generated when necessary
(separate equations are not generated if the static version is simply
equal to the dynamic version when lags and leads are removed).}

\item{latex}{A logical. If \code{TRUE} (the default), then LaTeX files
are created if the model block contains a \code{write_latex_static_model},
\code{write_latex_dynamic_model} or \code{write_latex_original_model} statement.}

\item{latex_options}{a list with options for writing LaTeX files.
See Details.}

\item{nostrict}{Obsolete: the logical negation of argument  \code{strict}. This
argument should not be used in new code: use argument \code{strict} instead.}

\item{silent}{A logical (default \code{FALSE}). If \code{TRUE}, then
output of the Dynare parser is suppressed except for warnings.}
}
\value{
A \code{DynMdl} object.
}
\description{
Creates a \code{\link{DynMdl}} object from a mod file. If the mod file
contains a fit block, then the \code{DynMdl} object implements the fit
procedure (except if argument \code{fit} is \code{FALSE}).
}
\details{
\subsection{Evaluation of model equations}{

There are several methods available for evaluating the model equations and
the Jacobian. These methods can be specified with argument \code{calc}.
The possible methods are
\describe{
\item{\code{R}}{the model equations and Jacobian are evaluated using R functions .
This is very slow for large models. This method should therefore only be
used for small models.}
\item{\code{bytcode}}{the same as \code{R}, except that the R function is turned into byte code.
This is usually only slightly faster than using the \code{R} method}
\item{\code{dll}}{The model equations and Jacobian are evaluated using a
shared library created for this specific model.
Function \code{dyn_mdl} generates C code for evaluating the equations and the Jacobian,
and subsequently compiles the C code to create a shared library. The
evaluation of the equations and Jacobian is much faster than for
the \code{R} and \code{bytecode} methods. However, the compilation of the
C code can take a considerable amount time for large models.}
\item{\code{internal}}{This methods converts
the equations to internal byte code using reverse Polish notation.
The internal byte code is evaluated using compiled C++ code that is part
of package \code{dynmdl}.
For this method the evaluation of the equations and Jacobian
is as fast as for the \code{dll} method, but the compilation time
is much faster. However, the \code{internal} method does not yet support
all features of Dynare models. For example, it cannot handle
model-local variables and some built-in functions (the only supported
built-in functions are currently  \code{exp}, \code{log}, \code{sqrt}, \code{abs} and \code{sign}).}
}

If possible, use the \code{internal} method, because this method is faster than
the other methods for both compiling the model and for evaluating the equations
and Jacobian. The \code{internal} method is therefore the default method.
However, as explained above, the mod file may contain features not
yet supported for the \code{internal} method, in which case another
method must be selected.
}

\subsection{Latex options}{
When the mod file contains a \code{write_latex_static_model},
\code{write_latex_dynamic_model} or \code{write_latex_oiriginal_model}
statement, then the Dynare parser of package \code{dynmdl} generates
LaTeX files in directory \code{latex}. Argument \code{latex_options} can be
used to change the format of the LaTeX files.
It should be a named list containing one or more of
the following components:
\describe{
\item{\code{par_as_num}}{A logical. If \code{TRUE}, then the parameters
are written as numerical constants to the LaTeX file, using the numerical
values as specified in the mod file. The default is \code{FALSE}.}
\item{\code{ndigits}}{The number of significant digits used when parameters
as written as numerical values (default 4). This argument is only used
if \code{par_as_num} is \code{TRUE}. For example, if \code{ndigits} is 4,
then the number \eqn{\pi} is printed as \code{3.142},
the number 120.25 as \code{120.2}, and the number
10.1234 as \code{1.012e+05}}
}}
}
