% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_mdl.R
\name{dyn_mdl}
\alias{dyn_mdl}
\title{Creates a \code{\link{DynMdl}} or  \code{\link{FitMdl}} object from a mod file}
\usage{
dyn_mdl(mod_file, period, data, base_period = NULL, calc = c("R",
  "bytecode", "dll", "internal"), fit_mod_file, debug = FALSE, dll_dir,
  max_laglead_1 = FALSE, nostrict = FALSE, fit = TRUE)
}
\arguments{
\item{mod_file}{the name of the model file (including extension .mod)}

\item{period}{a \code{\link[regts]{period_range}} object specifying the
model period, i.e. the period range for which the model will be solved. Thus
this period range excludes the lag and lead period.}

\item{data}{the model data as a  \code{\link[regts]{regts}} object with column
names}

\item{base_period}{a \code{\link[regts]{period}} object specifying
the base period for the trends. This is used if the model has trend variables.
All trend variables will be equal to 1 at the base period.}

\item{calc}{Method used to evaluate the model equations. 
Possible values are \code{"R"}, \code{"bytecode"}, \code{"dll"} and 
\code{"internal"}. See details.}

\item{fit_mod_file}{the name of the generated fit mod file. If not specified,
then the fit mod file is destroyed after the model has been parsed.}

\item{debug}{If \code{TRUE}, then some print statements
are executed.}

\item{dll_dir}{the directory where the dynamically linked library is stored.
Primarily used for testing.
Only used if argument \code{use_dll} is \code{TRUE}.}

\item{max_laglead_1}{a logical indicating whether the model should be 
transformed internally to a model with a maximum lag and lead of 1.
The default is \code{FALSE}. This option has no effect if the maximum lag
and lead of the original model is 1. Set this argument to
\code{TRUE} if you want to analyse the stability of the steady state with 
method \code{\link{check}} for models with a maximum lag or lead larger than 1.}

\item{nostrict}{Allows Dynare to issue a warning and continue processing when
there are more endogenous variables than equations,
an undeclared symbol is assigned in initval or endval,
or exogenous variables were declared but not used in the model block.}

\item{fit}{a logical. If \code{TRUE}, then the function returns
a  \code{FitMdl} object if a fit block has been found in the mod file.
If \code{FALSE} then this function does not return a \code{FitMdl} object.}
}
\value{
an \code{DynMdl} object or, if the mod file contains a
fit block, a \code{\link{FitMdl}} object.
}
\description{
Creates a \code{\link{DynMdl}} object from a mod file. If the mod file
contains a fit block, then this function returns a \code{\link{FitMdl}} 
object, which is an extension of a \code{DynMdl} object.
}
