% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl.R
\docType{class}
\name{DynMdl}
\alias{DynMdl}
\title{An R6 class for a Dynare model}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} containing a macro-economic model,
}
\description{
An R6 class for a Dynare model
}
\section{Methods}{

\describe{

\item{\code{\link{get_max_lag}}}{Returns the maximum lag}

\item{\code{\link{get_max_lead}}}{Returns the maximum lead}

\item{\code{\link{get_endo_names}}}{Returns the names of the endogenous variables.}

\item{\code{\link{get_exo_names}}}{Returns the names of the exogenous variables.}

\item{\code{\link{set_labels}}}{Set labels for the model variables}

\item{\code{\link{get_labels}}}{Returns the labels of the model variables and
parameters}

\item{\code{\link{get_tex_names}}}{Returns the LaTeX names of the model
variables and parameters}

\item{\code{\link{get_par_names}}}{Returns the names of the parameters.}

\item{\code{\link{set_param}}}{Sets the parameters of the model.}

\item{\code{\link{set_param_values}}}{Sets the values of one or more model parameters.}

\item{\code{\link{get_param}}}{Returns the parameters of the model.}

\item{\code{\link{set_static_exos}}}{Sets the static values of
the exogenous variables used to compute the steady state.}

\item{\code{\link{set_static_exo_values}}}{Sets the values of one or more
static exogenous variables}

\item{\code{\link{get_static_exos}}}{Returns the static values of
the exogenous variables.}

\item{\code{\link{set_static_endos}}}{Sets the static values of
the endogenous variables.}

\item{\code{\link{get_static_endos}}}{Returns the static values of
the endogenous variables.}

\item{\code{\link{set_static_data}}}{Sets the static values of
the model variables.}

\item{\code{\link{get_static_data}}}{Returns the static values of
the model variables.}

\item{\code{\link{get_all_static_endos}}}{Returns the static values of
the endogenous model variables, including fit instruments and
lagrange multupliers.}

#\item{\code{\link{get_all_static_data}}}{Returns the static values of
the exogenous and endogenous model variables, including fit instruments and
lagrange multupliers.}

\item{\code{\link{init_data}}}{Initializes the model data}

\item{\code{\link{set_period}}}{Sets the model period}

\item{\code{\link{get_period}}}{Returns the model period}

\item{\code{\link{get_data_period}}}{Returns the model data period.}

\item{\code{\link{get_lag_period}}}{Returns the lag period.}

\item{\code{\link{get_lead_period}}}{Returns the lead period}

\item{\code{\link{set_endo_values}}}{Sets the values of endogenous model
variables}

\item{\code{\link{set_exo_values}}}{Sets the values of exogenous model
variables}

\item{\code{\link{set_data}}}{Transfer timeseries to the model data}

\item{\code{\link{change_endo_data}}}{Changes the values of endogenous model
variables by applying a function}

\item{\code{\link{change_exo_data}}}{Changes the values of exogenous model
variables by applying a function}

\item{\code{\link{change_data}}}{Changes the values of endogenous and
exogenous model variables by applying a function.}

\item{\code{\link{get_data}}}{Returns the model data, including exogenous end exogenous
variables, fit instruments and trend variables (but excluding Lagrange multipliers).}

\item{\code{\link{get_all_endo_data}}}{Returns all endogenous model variables,
including exogenous end exogenous
variables, fit instruments and trend variables (but excluding Lagrange multipliers).}

\item{\code{\link{get_all_data}}}{Returns all endogenous end exogenous model varieables,
including fit instruments and Lagrange multipliers}

\item{\code{\link{get_endo_data}}}{Returns the endogenous model data}

\item{\code{\link{get_exo_data}}}{Returns the exogenous model data}

\item{\code{\link{get_vars_pars}}}{Returns a list with model variables and
parameters}

\item{\code{\link{solve_steady}}}{Solves the steady state}

\item{\code{\link{solve}}}{Solves the model}

\item{\code{\link{solve_steady_dynare}}}{Solves the steady state and eigenvalues with
Dynare (emplpying Matlab or Octave)}

\item{\code{\link{solve_dynare}}}{Solves the model with Dynare (employing Octave or Matlab)}

\item{\code{\link{check}}}{Compute the eigenvalues of the linear
system and check if the Blachard and Kahn conditions are satisfied.}

\item{\code{\link{residual_check}}}{Calculates the residuals of the equations
and reports the differences larger than a tolerance parameters}

\item{\code{\link{static_residual_check}}}{Calculates the residuals of the
static model equations and reports the differences larger than a tolerance parameters}

\item{\code{solve_perturbation}}{Solves the model using the perturbation
theory used in the Dynare function stoch_simul. Only shocks in the first
solution period are allowed.}

\item{\code{\link{get_jacob}}}{Returns the Jacobian for the dynamic model}

\item{\code{\link{get_static_jacob}}}{Returns the Jacobian for the
static version of the model}

\item{\code{\link{get_back_jacob}}}{Returns the Jacobian for a
backward looking model at a specific period}

\item{\code{\link{get_eigval}}}{Returns the eigenvalues computed with
method \code{\link{check}} or \code{solve_perturbation}}

\item{\code{\link{get_equations}}}{Returns a character vector with the
parsed equations of the model.}
\item{\code{\link{get_original_equations}}}{Returns a character vector with the
equations of the original model.}

\item{\code{\link{copy}}}{Returns a deep copy of the \code{\link{DynMdl}}
object}

\item{\code{\link{get_solve_status}}}{Returns the status of the last model
solve attempt}

\item{\code{\link{get_mdldef}}}{Returns a list with technical details of the
model.}
}
}

\section{Methods for the fit method}{

\describe{

\item{\code{\link{get_instrument_names}}}{Returns the names of the fit
instruments.}

\item{\code{\link{get_sigma_names}}}{Returns the names of the sigma
parameters used in the fit procedure.}

\item{\code{\link{set_fit_values}}}{Sets the values of the fit targets}

\item{\code{\link{set_fit}}}{Sets the targets for the fit procedure}

\item{\code{\link{set_fit_steady}}}{Sets the targets for the fit procedure
for the steady state.}

\item{\code{\link{get_fit}}}{Returns the fit targets
used in the fit procedure}

\item{\code{\link{get_fit}_steady}}{Returns the fit targets
used in the fit procedure for the steady state.}

\item{\code{\link{get_fit_instruments}}}{Returns all non-zero fit instruments
used in the fit procedure}

\item{\code{\link{set_sigma}}}{Sets one or more sigma parameters used
in the fit method}

\item{\code{\link{set_sigma_values}}}{Sets the values of sigma parameters
used in the fit method}

\item{\code{\link{get_sigmas}}}{Returns all sigma parameters >= 0 used in
the fit procedure. If a sigma parameter is negative, then the
corresponding fit instrument is not included}

\item{\code{\link{get_lagrange}}}{Returns the Lagrange multipliers
used in the fit procedure.}

\item{\code{\link{run_initval}}}{Run the initval equations to obtain
new values for endogenous and exogenous static variables for the current
values or parameters.}
}
}

\keyword{data}
