% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMod.R
\docType{class}
\name{DynMod}
\alias{DynMod}
\title{An R6 class for a Dynare model}
\format{\code{\link{R6Class}} object.}
\usage{
DynMod
}
\value{
Object of \code{\link{R6Class}} containing a macro-economic model,
}
\description{
An R6 class for a Dynare model
}
\section{Methods}{

\describe{

\item{\code{set_period(period)}}{Sets the model period. \code{period}
is a \code{\link{regperiod_range}} object or an object that can be coerced
to a \code{regperiod_range}. The model period is the longest period for which
the model may be solved. This method also allocates storage for
all model timeseries.  Model timeseries are
available for the so called 'model data period', which is
the model period extended with a lag and lead period. This method
also initialises all model timeseries with steady state values stored in the
\code{endos} and \code{exos} fields.}

\item{\code{solve_steady()}}{Solve the steady state of the model.
This methods solves the steady state problem. The data in field \code{endos}
is used as initial values for the endogenous variables.
This field is overwritten with the computed steady state values}

\item{\code{check()}}{Compute the eigenvalues of the linear
system and check if the Blachard and Kahn conditions are satisfied.}

\item{\code{solve(control = list())}}{Solves the model using a stacked-time
Newton method for the whole model period.
Argument \code{control} is a list with solve options (TODO: describe these
options somewhere).}

\item{\code{solve_perturbation()}}{Solves the model using the perturbation
theory used in the Dynare function stoch_simul. Only shocks in the first
solution period are allowed.}

\item{\code{get_jacob(sparse = TRUE)}}{Returns the Jacobian for the
stacked-time Newton problem either as a sparse matrix
(a \code{\link[Matrix]{Matrix}} object) or normal \code{\link{matrix}}.}

}
}
\keyword{data}

