% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMdl.R
\docType{class}
\name{FitMdl}
\alias{FitMdl}
\title{An R6 class for a Dynare model with Fit targets}
\format{\code{\link{R6Class}} object.}
\usage{
FitMdl
}
\value{
Object of \code{\link{R6Class}} containing a macro-economic model,
}
\description{
This class is a subclass of a \code{\link{DynMdl}} objects. It contains
special methods for the fit procedure.
}
\section{Methods}{

\describe{

\item{\code{get_endo_names}}{Returns the names of the endogenous variables
of the model (excluding the instruments and Lagrange multipliers used in the
fit procedure).}

\item{\code{get_exo_names}}{Returns the names of the exogenous variables
of the model (excluding the fit control variables).}

\item{\code{\link{get_instrument_names}}}{Returns the names of the fit 
instruments.}

\item{\code{\link{get_sigma_names}}}{Returns the names of the sigma
parameters used in the fit procedure.} 

\item{\code{\link{set_fit_values}}}{Sets the values of the fit targets}

\item{\code{\link{set_fit}}}{Sets the targets for the fit procedure} 

\item{\code{\link{get_fit}}}{Returns the fit targets
used in the fit procedure}

\item{\code{\link{clear_fit}}}{Remove all fit targets and set all fit sigmas 
to zero. Also all Lagrange multipliers are set to zero}

\item{\code{get_fit_instruments}}{Returns all non-zero fit instruments
used in the fit procedure}

\item{\code{get_sigmas}}{Returns all sigma parameters >= 0(rms values)
used in the fit procedure. If a sigma parameter is negative, then the 
corresponding fit instrument is not included}

\item{\code{get_lagrange}}{Returns the Lagrange multipliers
used in the fit procedure.}

\item{\code{get_exo_data(names, period = self$get_data_period()}}{Returns
the exogenous data excluding the control variables of the
fit procedure.
\code{pattern} is a regular expression,  \code{names} a list of variables
 and \code{period} an \code{\link[regts]{period_range}} object
 or an object that can be coerced to \code{period_range}.}

\item{\code{get_endo_data(pattern, names, period = self$get_data_period()}}{Returns
the endogenous data excluding the auxiliary endogenous variables
 for the fit procedure (the instruments and the Lagrange multipliers).
\code{pattern} is a regular expression,  \code{names} a list of variables
 and \code{period} an \code{\link[regts]{period_range}} object
 or an object that can be coerced to \code{period_range}.}
}
}

\keyword{data}
