% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMdl.R
\docType{class}
\name{FitMdl}
\alias{FitMdl}
\title{An R6 class for a Dynare model with Fit targets}
\format{\code{\link{R6Class}} object.}
\usage{
FitMdl
}
\value{
Object of \code{\link{R6Class}} containing a macro-economic model,
}
\description{
This class is a subclass of a \code{\link{DynMdl}} objects. It contains
special methods for the fit procedure.
}
\section{Methods}{

\describe{

\item{\code{get_endo_names()}}{Returns the names of the endogenous variables
of the model (excluding the residuals and Lagrange multipliers used in the
fit procedure).}

\item{\code{get_exo_names()}}{Returns the names of the exogenous variables
of the model (excluding the fit control variables).}

\item{\code{\link{set_fit_values}}}{Sets the values of the fit targets}

\item{\code{\link{set_fit}}}{Sets the targets for the fit procedure} 

\item{\code{\link{get_fit}}}{Returns the fit targets
used in the fit procedure}

\item{\code{\link{clear_fit}}}{Remove allfit targets}

\item{\code{get_residuals())}}{Returns the residuals used in the 
fit procedure.}

\item{\code{get_lagrange())}}{Returns the Lagrange multipliers
used in the fit procedure.}

\item{\code{get_exo_data(names, period = self$get_data_period()}}{Returns
the exogenous data excluding the control variables of the
fit procedure.
\code{pattern} is a regular expression,  \code{names} a list of variables
 and \code{period} an \code{\link[regts]{period_range}} object
 or an object that can be coerced to \code{period_range}.}

\item{\code{get_endo_data(pattern, names, period = self$get_data_period()}}{Returns
the endogenous data excluding the auxiliary endogenous variables
 for the fit procedure (the residuals and the Lagrange multipliers).
\code{pattern} is a regular expression,  \code{names} a list of variables
 and \code{period} an \code{\link[regts]{period_range}} object
 or an object that can be coerced to \code{period_range}.}
}
}

\keyword{data}
