% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_data-methods}
\alias{get_data-methods}
\alias{get_data}
\alias{get_endo_data}
\alias{get_exo_data}
\title{\code{\link{DynMdl}} methods: Retrieve timeseries from the model data}
\description{
These methods of R6 class \code{\link{DynMdl}} 
can be used to retrieve timeseries from the model data.

If the \code{DynMdl} object is also a \code{\link{FitMdl}} object, then
the \code{get-data} methods do not include the the auxiliary endogenous and
exogenous variables used in the fit procedure. Use 
\code{\link{get_fit}}, \code{\link{get_fit_instruments}} and 
\code{\link{get_lagrange}} to obtain the various variables used 
in the fit procedure.
}
\section{Usage}{

\preformatted{

mdl$get_data(pattern, names, period = mdl$get_data_period())

mdl$get_endo_data(pattern, names, period = mdl$get_data_period())

mdl$get_exo_data(pattern, names, period = mdl$get_data_period())

}

\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{names}}{a character vector with variable names}
\item{\code{pattern}}{a regular expression}
\item{\code{period}}{an \code{\link[regts]{period_range}} object or an
object that can be coerced to a \code{period_range}}
}

If neither \code{names} nor \code{pattern} have been specified,
then all endogenous or exogenous variables are set to the specified value.
}

\section{Methods}{

\itemize{

\item \code{get_data}: All model variables
\item \code{get_endo_data}: Endogenous model variables

\item \code{get_exo_data}: Exogenous model variables
}
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2017Q3")

mdl$get_data(names = "c", pattern  = "y.", period = "2017Q1/2017Q2")
}
\seealso{
\code{\link{get_fit}}, \code{\link{get_fit_instruments}},
\code{\link{get_lagrange}} and \code{\link{get_vars_pars}}.
}
