% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_data-methods}
\alias{get_data-methods}
\alias{get_data}
\alias{get_endo_data}
\alias{get_exo_data}
\alias{get_trend_data,}
\alias{get_all_endo_data}
\alias{get_all_exo_data}
\alias{get_all_data}
\title{\code{\link{DynMdl}} methods: Retrieve timeseries from the model data}
\description{
These methods of R6 class \code{\link{DynMdl}}
can be used to retrieve timeseries from the model data.

If the \code{DynMdl} object is also a \code{\link{DynMdl}} object, then
\code{get_data} also returns the fit instruments. In contrast,
\code{get_endo_data} does not return these fit instruments.
Both \code{get_data} and \code{get_endo_data} do not return
the Lagrange multipliers used in the fit method. Use method
\code{\link{get_lagrange}} to obtain these Lagrange multipliers.
\code{\link{get_all_endo_data}} returns all endogenous variables,
fit instruments and Lagrange multipliers, for the complete data period.
\code{\link{get_all_data}} also returns all exogenous variables.
These functions can be useful to save the complete solution of a model that
be used as initial values for the endogenous variables in another
model.
}
\section{Usage}{

\preformatted{

mdl$get_data(pattern, names, period = mdl$get_data_period(), 
             trend = TRUE)

mdl$get_endo_data(pattern, names, period = mdl$get_data_period(),
                  trend = TRUE)

mdl$get_exo_data(pattern, names, period = mdl$get_data_period())

mdl$get_trend_data(pattern, names, period = mdl$get_data_period())

mdl$get_all_endo_data(trend = TRUE)

mdl$get_all_data(trend = TRUE)

}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression}
\item{\code{names}}{a character vector with variable names}
\item{\code{period}}{an \code{\link[regts]{period_range}} object or an
object that can be coerced to a \code{period_range}}
\item{\code{trend}}{a logical. This argument is used for model with
trend variables. If \code{TRUE} (the default), then the
endogenous variables are multiplied with their trends (called deflators
in the mod file)}
}

If neither \code{names} nor \code{pattern} have been specified,
then all variables with the specific type are returned.
}

\section{Methods}{

\itemize{

\item \code{get_data}: All model variables: exogenous and endogenous model
variables, trends variables, and fit instruments for \code{\link{DynMdl}}
objects
\item \code{get_endo_data}: Endogenous model variables, excluding
fit instruments.
\item \code{get_exo_data}: Exogenous model variables
\item \code{get_trend_data}: Trend variables (variables declared with
\code{trend_var} in the mod file).
\item \code{get_all_endo_data}: All endogenous variables, including
fit instruments and lagrange multipliers.
\item \code{get_all_data}: All endogenous and exogenous variables, including
fit instruments and lagrange multipliers.
}
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2017Q3")

mdl$get_data(names = "c", pattern  = "y.", period = "2017Q1/2017Q2")
}
\seealso{
\code{\link{get_fit-methods}}, \code{\link{get_fit}},
\code{\link{get_fit_instruments}}, \code{\link{get_lagrange}} and
\code{\link{get_vars_pars}}.
}
