% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMod_doc.R
\name{get_data_period}
\alias{get_data_period}
\title{\code{\link{DynMod}} method: returns the model data period}
\description{
This method of R6 class \code{\link{DynMod}} returns the model data period.
}
\section{Usage}{

\preformatted{
mdl$get_data_period()

}

\code{mdl} is an \code{DynMod} object
}

\seealso{
\code{\link{set_period}}
}
