% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_equations}
\alias{get_equations}
\alias{get_original_equations}
\title{\code{\link{DynMdl}} method: Returns a character vector with the model
equations.}
\description{
These method of R6 class \code{\link{DynMdl}} returns a character vector
with the model equations (excluding local equations). \code{get_equations}
returned the parsed equations, where trends (if present) have been removed.
\code{get_original_equations} returns the equations as defined in the mod
file.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
 md$get_equations(i)
 mdl$get_original_equations(i)
}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{i}}{A numeric vector with the indices of the non-local equations.
If not specified, then the function returns all equations}
}
}

\examples{
mdl <- islm_mdl(period = "2018Q1/2023Q3")

# print the 4th equation nicely to the screen
cat(mdl$get_equations(4))

# print all equations
print(mdl$get_equations())
}
