% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{get_name-methods}
\alias{get_name-methods}
\alias{get_endo_names}
\alias{get_exo_names}
\alias{get_par_names}
\title{\code{\link{DynMdl}} methods: Retrieve the names of model variables or
parameters}
\description{
These methods of R6 class \code{\link{DynMdl}} 
return the names of the model variables or parameters

If the \code{DynMdl} object is also a \code{\link{DynMdl}} object, then
\code{get_endo_names} and \code{get_exo_names} do not include the names of 
the auxiliary endogenous and exogenous variables used in the fit procedure.
Use \code{\link{get_instrument_names}} to obtain the names of the fit 
instruments.
}
\section{Usage}{

\preformatted{
mdl$get_endo_names(type = c("all", "lags", "leads")

mdl$get_exo_names()

mdl$get_par_names()

}

\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{type}}{a character describing the type of the endogenous variables:
 \code{"lags"} or \code{"leads"} for endogenous variables with lags or leads,
respectively. The default is \code{"all"} (all endogenous variables).}
}
}

\section{Methods}{

\itemize{
\item \code{get_endo_names}: Names of the endogenous model variables

\item \code{get_exo_names}: Names of the endogenous model variables

\item \code{get_par_names}: Names of the model parameters

}
}

\seealso{
\code{\link{get_instrument_names}} and \code{\link{get_sigma_names}}
}
