% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{init_data}
\alias{init_data}
\title{\code{\link{DynMdl}} method: initializes the model data.}
\description{
This method of R6 class \code{\link{DynMdl}} initializes the
model data.

This method sets the model data period and initializes
the model variables with static values
of the exogenous and endogenous model variables.

This methods also sets the model period, the standard period
for which the model will be solved. The model period
is obtained from the data period by subtracting the lag and lead periods.
}
\section{Usage}{

\preformatted{
mdl$init_data(data_period)

}

\code{mdl} is a \code{DynMdl} object
}

\section{Arguments}{


\describe{
\item{\code{data_period}}{\code{\link[regts]{regperiod_range}}
object, or an object that can be coerced to 
\code{\link[regts]{regperiod_range}}}
}
}

