% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mdl.R
\name{read_mdl}
\alias{read_mdl}
\title{Reads a model from a RDS file}
\usage{
read_mdl(file, dll_dir, silent = FALSE)
}
\arguments{
\item{file}{the name of the RDS file}

\item{dll_dir}{the directory where the dynamically linked library is stored.
Primarily used for testing. Only used if the model
was created with the dll option (see function  \code{\link{dyn_mdl}}).}

\item{silent}{A logical (default \code{FALSE}). It \code{TRUE}, then no
output is written.}
}
\value{
a \code{\link{DynMdl}} object.
}
\description{
This function reads a model from an RDS file that has been written
by method \code{\link{write_mdl}} of an \code{\link{DynMdl}} object.
}
\examples{
mdl <- islm_mdl("2017Q1/2019Q2")
mdl$write_mdl("islm_mod.rds")
mdl2 <- read_mdl("islm_mod.rds")
}
\seealso{
\code{\link{write_mdl}}
}
