% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{residual_check}
\alias{residual_check}
\title{\code{\link{DynMdl}} method: Calculates the residuals of the equations}
\description{
This method of R6 class \code{\link{DynMdl}} calculates the 
residuals for the full model period and returns the 
result  as a \code{\link[regts]{regts}} timeseries object.
}
\section{Usage}{

\preformatted{
mdl$residual_check(tol, include_fit_eqs = FALSE)

}

\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{tol}}{the tolerance parameter.
If specified, then the return value does not include columns for the equations
whose residuals are smaller than \code{tol}}
\item{\code{include_fit_eqs}}{a logical value (default \code{FALSE}).  
This argument is only used if \code{mdl} is a \code{\link{DynMdl}} object. 
If \code{TRUE}, then the fit equations are included in the residual check.}
\item{\code{debug_eqs}}{Debug equations (default \code{FALSE}). Only used
for the internal calculation mode (\code{calc == "internal"},
see \code{\link{dyn_mdl}}). If
\code{TRUE} then numerical problems in evaluation
of mathematical functions or operators such a \code{log} are reported.}
}
}

\seealso{
\code{\link{static_residual_check}}
}
