% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{set/get_param}
\alias{set/get_param}
\alias{set_param}
\alias{set_param_values}
\alias{get_param}
\title{\code{\link{DynMdl}} methods: Set and get model parameters}
\description{
\code{\link{DynMdl}} methods \code{set_param} and \code{set_param_values}
can be used to set the model parameters. Method \code{get_param}
returns the model parameters, including the sigma parameters used in
the fit method.
}
\section{Usage}{

\preformatted{
mdl$set_param(params, names, name_err = c("stop", "warn", "silent"))

mdl$set_param_values(value, names, pattern)

mdl$get_param(pattern, names)

}
\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{params}}{a named numeric vector with parameter values.
The names are the names of the parameters.}
\item{\code{names}}{a character vector with names of the parameters.
For method \code{set_param}, this argument \emph{must} be specified if
\code{params} is a vector without names.}
\item{\code{name_err}}{this option specifies the action that should be taken
when a specified name is not the name of a model parameter.
For \code{"stop"} (the default), the execution of this function is stopped.
For \code{"warn"} and \code{"silent"} the names that are no model
parameters are skipped. \code{"warn"} does, however, give a warning.}
\item{\code{pattern}}{a regular expression. The action (get or
set parameter values) is applied to all parameters wiht names
matching \code{pattern}.}
}

If neither \code{names} nor \code{pattern} have been specified in
methods \code{set_param_values} or \code{get_param}, then the action
is applied to all model parameters.
}

\section{Methods}{

\itemize{
\item \code{set_param}: Set the parameters using a named numeric
vector. The names of the vector should be the parameter names.
\item \code{set_param_values}: Give one or more parameter
a specified value.
\item \code{get_param}: Returns the parameters.
}
}

\examples{
mdl <- islm_mdl()

# print parameters c0, c1, c2 and c3
print(mdl$get_param(pattern = "^c.*"))

# set parameter i0 to 101 and c0 to 110
mdl$set_param(c(i0 = 101, c0 = 110))

# set all parameters with names ending with "5" to zero:
mdl$set_param_values(0, pattern = "5$")

# print all parameters
print(mdl$get_param())
}
\seealso{
\code{\link{get_par_names}}, \code{\link{set_sigma}},
\code{\link{set_sigma_values}} and \code{\link{get_sigmas}}
}
