% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{set/get_static_endos/exos}
\alias{set/get_static_endos/exos}
\alias{set_static_exos}
\alias{set_static_endos}
\alias{set_static_data}
\alias{get_static_endos}
\alias{get_static_exos}
\alias{get_static_data}
\alias{get_all_static_endos}
\alias{get_all_static_data}
\title{\code{\link{DynMdl}} methods: set and get the static values of the
model variables}
\description{
\code{set_static_exos},  \code{set_static_endos} and  \code{set_static_data}
can be used to set one or
more static values of the endogenous and/or exogenous model variables.
The corresponding \code{get} methods can be used to retrieve them.

Each \code{\link{DynMdl}} object contains a set of static values
for the exogenous and endogenous model variables.
The static exogenous values are used to compute the steady state
with function methode \code{\link{solve_steady}}.
The static endogenous values are both input and output of
\code{solve_steady}: they are used as an initial guess
for the steady state, and replaced by the steady
state solution.

The static values are initialized to the values specified in the initval block of
the mod file, or to zero if they are not specified in the initval block.
The static values can be modified with methods \code{set_static_endos},
\code{set_static_endos},  \code{set_static_data} or
\code{set_static_exo_values}.

\code{\link{get_all_static_endos}} returns the static values for all
endogenous variables, fit instruments and Lagrange multipliers.
\code{\link{get_all_data}} also returns all static values of exogenous variables.
These functions can be useful to save the complete solution of a model that
be used as initial values for the endogenous variables in another
model.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
mdl$set_static_endos(endos, names = names(endos), name_err = c("stop", "warn", "silent"))
mdl$set_static_exos(exos, names = names(endos), name_err = c("stop", "warn", "silent"))
mdl$set_static_data(data, names = names(data), name_err = c("stop", "warn", "silent"))
mdl$get_static_endos(pattern, names)
mdl$get_static_endos(pattern, names)
mdl$get_static_data(pattern, names)
mdl$get_all_static_endos()
mdl$get_all_static_data()
}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{

\describe{
\item{\code{endos}}{A (named) numerical vector with new static values of the
endogenous variables. If the vector has no names, than argument
\code{names} must be specified.}
\item{\code{exos}}{A (named) numerical vector with new static values of the
exogenous variables. If the vector has no names, than argument
\code{names} must be specified.}
\item{\code{data}}{A (named) numerical vector with new static values of
both endogenous and exogenous variables.  If the vector has no names, than argument
\code{names} must be specified.}
\item{\code{names}}{a character vector with names of model variables}
\item{\code{pattern}}{a regular expression}
\item{\code{name_err}}{this option specifies the action that should be taken
when a variable name is not a model variable.
For \code{"stop"} (the default), the execution of this function is stopped.
For \code{"warn"} and \code{"silent"} the names that are no model
variables are skipped. \code{"warn"} does however give a warning.}
}
}

\section{Methods}{

\itemize{
\item \code{set_static_endos}: Set the static values of one or more
endogenous variables (including static fit instruments and static Lagrange
multipliers).
\item \code{set_static_exos}: Set the static values of one or more
exogenous variables.
\item \code{set_static_data}: Set the static values of one or more
endogenous or exogenous variable
(including  static fit instruments and static Lagrange multipliers.)
\item \code{get_static_endos}: Returns the static values of one or more
endogenous variables, excluding fit instruments and Lagrange multipliers.
\item \code{get_static_exos}: Returns the static values of one or more
exogenous variables.
\item \code{get_static_data}: Returns the static values of the model
variables: exogenous and endogenous model variables and static fit instruments.
\item \code{get_all_static_endos}: Returns all static endogenous variables,
including fit instruments and lagrange multipliers.
\item \code{get_all_static_data}: Returns all static endogenous and exogenous variables,
including fit instruments and lagrange multipliers.
}
}

\examples{
mdl <- islm_mdl()
mdl$set_static_endos(c(y = 1250))

#  set static values of all exogenous variables starting with m 
# (for this model only "ms") to zero. 
mdl$set_static_exo_values(333, pattern = "^m")

print(mdl$get_static_endos())
}
\seealso{
\code{\link{set_static_values-methods}},
\code{\link{change_static_data-methods}}, \code{\link{solve_steady}},
\code{\link{check}}
}
