% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{set_param}
\alias{set_param}
\title{\code{\link{DynMdl}} method: Sets the model parameters}
\description{
This method of R6 class \code{\link{DynMdl}}
sets the model parameters
}
\section{Usage}{

\preformatted{
mdl$set_param(p, name_err = c("stop", "warn", "silent"))

}
\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{p}}{a named numeric vector with parameter values.
The names are the names of the parameters}
\item{\code{name_err}}{this option specifies the action that should be taken 
when a variable name is not a model variable.
For \code{"stop"} (the default), the execution of this function is stopped.
For \code{"warn"} and \code{"silent"} the names imeseries that are no model
parameters are skipped. \code{"warn"} does however give a warning.}
}
}

\examples{
mdl <- islm_mdl()
mdl$set_param(c(i0 = 101))
}
\seealso{
\code{\link{set_param_values}} and \code{\link{get_param}}
}
