% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{set_values-methods}
\alias{set_values-methods}
\alias{set_endo_values}
\alias{set_exo_values}
\title{\code{\link{DynMdl}} methods: Sets the values of the model data}
\description{
This method of R6 class \code{\link{DynMdl}} 
can be used to set the values of the model data.
}
\section{Usage}{

\preformatted{

mdl$set_endo_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_exo_values(value, names, pattern, period = mdl$get_data_period())

}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{value}}{a numeric vector of length 1 or with the same length
as the length of the range of \code{period}. 
If the model has trends, then the values should 
include the trends.}
\item{\code{names}}{a character vector with variable names}
\item{\code{pattern}}{a regular expression}
\item{\code{period}}{a \code{\link[regts]{period_range}} object or an
object that can be coerced to a \code{period_range}}

If neither \code{names} nor \code{pattern} have been specified,
then all endogenous or exogenous variables are set to the specified value.
}
}

\section{Methods}{

\itemize{
\item \code{set_endo_values}: Endogenous model variables

\item \code{set_exo_values}: Exogenous model variables

}
}

\examples{

mdl <- islm_mdl(period = "2017Q1/2018Q3")

# set the values of ms in 2017Q1 and 2017Q2
mdl$set_exo_values(c(205, 206), names = "ms", period = "2017Q1/2017Q2")

# set the values for y and yd to 1000 for  the full data period
mdl$set_endo_values(1000, pattern = "^yd?$")
}
\seealso{
\code{\link{change_data-methods}} and  \code{\link{set_data}}
}
