% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{solve}
\alias{solve}
\title{\code{\link{DynMdl}} method: Solves the model}
\description{
This method of R6 class \code{\link{DynMdl}} solves
the model.

\code{solve} does \emph{not} raise an error when the solve was
not successful. In that case a warning may be issued.
Method \code{\link{get_solve_status}} can be used to check
whether the solve was successfully terminated or not.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
 md$solve(control = list(), mode, solver = c("umfpackr", "nleqslv"), 
          start = c("current", "previous"), debug_eqs = FALSE,
          homotopy = FALSE, ...)
}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{control}}{A named list of control parameters passed to function
\code{\link[umfpackr]{umf_solve_nl}} or \code{\link[nleqslv]{nleqslv}},
depending on argument \code{solver}}
\item{\code{mode}}{A character specifying the solve mode. Possible values
are \code{"stacked_time"} and \code{"backwards"}. By default, models
with leads are solved with the stacked time method and models without leads
are solved backwards.}
\item{\code{solver}}{Specifies the solver employed to solve the model:
\code{umfpackr} (sparse linear algebra) or \code{nleqslv} (dense linear algebra).
For large model, the \code{umfpackr} solve can be much faster.}
\item{\code{start}}{Method used to initialize starting values when solving
the model backwards. For \code{"current"} (the default) the current values
of the endogenous variables are used as starting values. For
\code{"previous"} the solution of the previous period is used to create
starting values (except for the first period when the model is solved).
This argument is ignored if the model if solved with the stacked time Newton
method}
\item{\code{debug_eqs}}{Debug equations (default \code{FALSE}). Only used
for the internal calculation mode (\code{calc == "internal"},
see \code{\link{dyn_mdl}}). If
\code{TRUE} then numerical problems in evaluation
of mathematical functions or operators such a \code{log} are reported.}
\item{\code{homotopy}}{A logical. If \code{TRUE} (the default), then the
homotopy approach is used when directly solving the model fails.
Consult the documentation of Dynare for more information about the homotopy aproach.}
\item{\code{...}}{Other arguments passed to the solver}
}
}

\examples{
islm <- islm_mdl(period = "2018Q1/2023Q3")
islm$set_exo_values(260, period = "2018q1", names = "g")
islm$solve(control = list(trace = TRUE))
}
\seealso{
\code{\link{solve_steady}} and \code{\link{get_solve_status}}
}
