% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{solve_dynare}
\alias{solve_dynare}
\title{\code{\link{DynMdl}} method: Solves the model with Dynare}
\description{
Solve the model with Dynare using Matlab or Octave.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
mdl$solve_dynare(scratch_dir = tempfile(), 
                 dynarere_path = NULL, model_options = list(), 
                 solve_options = list(tolf = 1e-8, tolx = 1e-8)),
                 use_octave = Sys.which("matlab") == "") 
}

\code{mdl} is a \code{\link{DynMdl}} object
}

\section{Arguments}{

\describe{
\item{\code{scratch_dir}}{Directory where the Matlablab and Dynare scripts are 
created.  By default this is a temporary directory that is automatically
deleted when the R session terminates.}
\item{\code{dynare_path}}{Character string specifying the name of the 
directory of the Dynare installation. On Linux it is usually not necessary 
to the specify this argument. On Windows it is necessary to specify the path 
of the Dynare installation. In you are running R in the CPB 
environment the path to Dynare is set automatically.}
\item{\code{model_options}}{Options passed to the \code{model} command of
Dynare. This should be a named list, which names corresponding to the Dynare
options. Specify a \code{NULL} value if the option has no value.
Consult the documentation of  Dynare for a list of available options.
Example: \code{model_options = list(block = NULL, mfs = 2)}}
\item{\code{solve_options}}{Options passed to the 
\code{perfect_foresight_solver} command of
Dynare. This should be a named list, which names corresponding to the Dynare
options. Specify a \code{NULL} value if the option has no value.
Consult the documentation of  Dynare for a list of available options.
Example: \code{steady_options = list(tolf = 1e-7, no_homotopy = NULL)}.
The default options are list(tolf = 1e-8, tolx = 1e-8)}
#' \item{\code{use_octave}}{A logical. If \code{TRUE}, then
Dynare is envoked with Octave, otherwise Matlab is used. By default 
Matlab is used if available.}
}
}

\examples{
\dontrun{
 islm <- islm_mdl(period = "2018Q1/2023Q3")
 islm$set_exo_values(260, period = "2018q1", names = "g")
 islm$solve_dynare()
}
}
