% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{solve_steady}
\alias{solve_steady}
\title{\code{\link{DynMdl}} method: Solves the steady state.}
\description{
This method of R6 class \code{\link{DynMdl}} solves the steady state.

This function uses the static exogenous and endogenous variables
stored in the \code{DynMdl} object. The static endogenous variables
are used as an initial guess for solving the steady state.
After creating a \code{DynMdl} object, the static exogenous and endogenous
variables are initialized to the values specified in the initval block of 
the mod file, or to zero if they are not specified in the initval block.
The static variables can be modified with methods
\code{\link{set_static_exos}} and \code{\link{set_static_endos}}.

The function \code{\link{get_static_endos}} can be used to retrieve the
steady state solution.

\code{solve_steady} does \emph{not} raise an error when the solve was
not successful. In that case a warning may be issued. 
Method \code{\link{get_solve_status}} can be used to check
whether the solve was successfully terminated or not.
}
\section{Usage}{

\code{DynMdl} method:
\preformatted{
mdl$solve_steady(control, solver = c("umfpackr", "nleqslv"), ...)

}

\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{control}}{A named list of control parameters passed to function
\code{\link[umfpackr]{umf_solve_nl}} or \code{\link[nleqslv]{nleqslv}},
depending on argument \code{solver}}
\item{\code{solver}}{Specifies the solver employed to solve the model:
\code{umfpackr} (sparse linear algebra) or \code{nleqslv} (dense linear algebra).
For large model, the \code{umfpackr} solve can be much faster.}
\item{\code{...}}{Other arguments passed to the solver}
}
}

\examples{
mdl <- islm_mdl(period = "2018Q1/2080Q1")
mdl$solve_steady(control = list(trace = 1))

# print the solution
print(mdl$get_static_endos())

# update the model data with steady state values of endogenous variables
mdl$put_static_endos()
}
\seealso{
\code{\link{set_static_endos}}, \code{\link{set_static_exos}},
\code{\link{get_static_endos}}, \code{\link{get_static_exos}},
\code{\link{put_static_endos}} and \code{\link{get_solve_status}}
}
