% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMdl_doc.R
\name{static_residual_check}
\alias{static_residual_check}
\title{\code{\link{DynMdl}} method: Calculates the residuals of the equations
of the static model}
\description{
This method of R6 class \code{\link{DynMdl}} calculates the 
residuals for the static version of the model. The result is a named numeric
vector, where the names are the equation numbers.
}
\section{Usage}{

\preformatted{
mdl$static_residual_check(tol, include_fit_eqs = FALSE)

}

\code{mdl} is an \code{\link{DynMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{tol}}{the tolerance parameter. 
If specified, then return value does not include equations whose residuals 
are smaller than \code{tol}}
 \item{\code{include_fit_eqs}}{a logical value (default \code{FALSE}).  
This argument is only used if \code{mdl} is a \code{\link{FitMdl}} object. 
If \code{TRUE}, then the fit equations are included in the residual check.}
}
}

\seealso{
\code{\link{residual_check}}
}
